+++ api/src/users/rider.spec.js
@@ -9,9 +9,9 @@ import Server from '../config/server';
 
 const url = `http://${api().host}:${api().port + api().version}`;
 const rider = {
-    lastname: 'Doe',
     firstname: 'John',
-    phone: '+33600000000',
+    lastname: 'Doe',
+    phone: '0600000000',
     email: 'john@doe.com',
     password: '123456'
 };
@@ -37,9 +37,9 @@ describe('POST /riders', () => {
 
     it('should return several errors', (done) => {
         chai.request(url).post('/riders').send({
-            lastname: 'Doe',
             firstname: 'John',
-            phone: '+3360000000',
+            lastname: 'Doe',
+            phone: '060000000',
             email: 'john@doe',
             password: '12345'
         }).end((err, res) => {
@@ -60,7 +60,7 @@ describe('POST /riders', () => {
         chai.request(url).post('/riders').send(rider).end((err, res) => {
             expect(res).to.have.status(201);
             expect(res).to.have.header('content-type', 'application/json; charset=utf-8');
-            expect(res).to.have.header('location', '/v1/authenticate/login');
+            expect(res).to.have.header('location', `/v1/riders/${res.body.data[0].rider.uuid}`);
             expect(res.body.success).equal(true);
             expect(res.body.status).equal(201);
             expect(res.body.message).to.be.a('string');
@@ -70,21 +70,7 @@ describe('POST /riders', () => {
         });
     });
 
-    it('should return error email already exists', (done) => {
-        chai.request(url).post('/riders').send(rider).end((err, res) => {
-            expect(res).to.have.status(400);
-            expect(res).to.have.header('content-type', 'application/json; charset=utf-8');
-            expect(res.body.success).equal(false);
-            expect(res.body.status).equal(400);
-            expect(res.body.errors).all.include.keys('message', 'code');
-            expect(res.body.errors).contain.a.thing.with.property('code', 'email_address_already_taken');
-
-            done();
-        });
-    });
-
-    it('should return error phone number already exists', (done) => {
-        rider.email = 'john2@doe.com';
+    it('should return error phone number already taken', (done) => {
         chai.request(url).post('/riders').send(rider).end((err, res) => {
             expect(res).to.have.status(400);
             expect(res).to.have.header('content-type', 'application/json; charset=utf-8');
@@ -92,6 +78,20 @@ describe('POST /riders', () => {
             expect(res.body.status).equal(400);
             expect(res.body.errors).all.include.keys('message', 'code');
             expect(res.body.errors).contain.a.thing.with.property('code', 'phone_number_already_taken');
+
+            done();
+        });
+    });
+
+    it('should return error email already taken', (done) => {
+        rider.phone = '0600000001';
+        chai.request(url).post('/riders').send(rider).end((err, res) => {
+            expect(res).to.have.status(400);
+            expect(res).to.have.header('content-type', 'application/json; charset=utf-8');
+            expect(res.body.success).equal(false);
+            expect(res.body.status).equal(400);
+            expect(res.body.errors).all.include.keys('message', 'code');
+            expect(res.body.errors).contain.a.thing.with.property('code', 'email_address_already_taken');
 
             done();
         });