+++ app/src/app/core/config.ts
@@ -1,0 +1,4 @@
+export const config = {
+    suffixTitle: ' - Uber Like',
+    baseUrl: 'http://localhost:1337/v1'
+};