+++ app/src/app/home/home.component.ts
@@ -1,0 +1,18 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'uberlike-home',
+    templateUrl: './home.component.html',
+    styleUrls: ['./home.component.scss']
+})
+export class HomeComponent implements OnInit {
+
+    constructor() { }
+
+    /**
+     * When component have been initialized
+     */
+    ngOnInit(): void {
+    }
+
+}