+++ app/src/app/users/rider-detail/rider-detail.component.ts
@@ -1,0 +1,18 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'uberlike-rider-detail',
+    templateUrl: 'rider-detail.component.html',
+    styleUrls: ['rider-detail.component.scss']
+})
+export class RiderDetailComponent implements OnInit {
+
+    constructor() { }
+
+    /**
+     * When component have been initialized
+     */
+    ngOnInit(): void {
+    }
+
+}