+++ app/src/app/users/rider.model.ts
@@ -1,0 +1,7 @@
+export class Rider {
+    firstname: string;
+    lastname: string;
+    phone: string;
+    email: string;
+    password: string;
+}